pandas-transformer/requirements.txt
pandas==2.2.3
google-cloud-storage==2.18.2
